refactor: just export isEqualBlock from handleHotUpdate.ts
